+++ __tests__/config/fixtures.ts
@@ -33,8 +33,6 @@ export const compiledL1L2 = readContract('l1l2_compiled');
 export const compiledTypeTransformation = readContract('contract');
 export const compiledMulticall = readContract('multicall');
 export const compiledTestDapp = readContract('TestDapp');
-export const compiledStarknetId = readContract('starknetId_compiled');
-export const compiledNamingContract = readContract('naming_compiled');
 export const compiledHashSierra = readContractSierra('cairo/hash/hash');
 export const compiledHashSierraCasm = readContractSierraCasm('cairo/hash/hash');
 export const compiledHelloSierra = readContractSierra('cairo/helloSierra/hello');
@@ -58,7 +56,17 @@ export const compiledDummy2Eth = readContractSierra('cairo/ethSigner/dummy2ForEt
 export const compiledDummy2EthCasm = readContractSierraCasm('cairo/ethSigner/dummy2ForEth');
 export const compiledEthPubk = readContractSierra('cairo/ethSigner/testEthPubKey.sierra');
 export const compiledEthPubkCasm = readContractSierraCasm('cairo/ethSigner/testEthPubKey');
-
+export const compiledTuple = readContractSierra('cairo/cairo253/tupleResponse.sierra');
+export const compiledTupleCasm = readContractSierraCasm('cairo/cairo253/tupleResponse');
+// StarknetId
+export const compiledStarknetId = readContractSierra('starknetId/identity/identity.sierra');
+export const compiledStarknetIdCasm = readContractSierraCasm('starknetId/identity/identity');
+export const compiledNaming = readContractSierra('starknetId/naming/naming.sierra');
+export const compiledNamingCasm = readContractSierraCasm('starknetId/naming/naming');
+export const compiledPricing = readContractSierra('starknetId/pricing/pricing.sierra');
+export const compiledPricingCasm = readContractSierraCasm('starknetId/pricing/pricing');
+export const compiledSidMulticall = readContractSierra('starknetId/multicall/multicall.sierra');
+export const compiledSidMulticallCasm = readContractSierraCasm('starknetId/multicall/multicall');
 export function getTestProvider(isProvider?: true): ProviderInterface;
 export function getTestProvider(isProvider?: false): RpcProvider;
 export function getTestProvider(isProvider: boolean = true): ProviderInterface | RpcProvider {