+++ src/provider/rpc.ts
@@ -1,4 +1,6 @@
-import { RpcChannel } from '../channel/rpc_0_6';
+import { ProviderInterface } from './interface';
+import { LibraryError } from './errors';
+import { RpcChannel, RPC06, RPC07 } from '../channel';
 import {
   AccountInvocations,
   BigNumberish,
@@ -27,12 +29,11 @@ import {
 import { getAbiContractVersion } from '../utils/calldata/cairo';
 import { isSierra } from '../utils/contract';
 import { RPCResponseParser } from '../utils/responseParser/rpc';
-import { ProviderInterface } from './interface';
 
 export class RpcProvider implements ProviderInterface {
   private responseParser = new RPCResponseParser();
 
-  public channel: RpcChannel;
+  public channel: RPC07.RpcChannel | RPC06.RpcChannel;
 
   constructor(optionsOrProvider?: RpcProviderOptions | ProviderInterface | RpcProvider) {
     if (optionsOrProvider && 'channel' in optionsOrProvider) {
@@ -93,6 +94,13 @@ export class RpcProvider implements ProviderInterface {
 
   public async getBlockWithTxs(blockIdentifier?: BlockIdentifier) {
     return this.channel.getBlockWithTxs(blockIdentifier);
+  }
+
+  public async getBlockWithReceipts(blockIdentifier?: BlockIdentifier) {
+    if (this.channel instanceof RPC06.RpcChannel)
+      throw new LibraryError('Unsupported method for RPC version');
+
+    return this.channel.getBlockWithReceipts(blockIdentifier);
   }
 
   public getStateUpdate = this.getBlockStateUpdate;